tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/022.html

<!DOCTYPE html>
<html><head>
        <title> scheduler: DOM added script, late .src </title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="testlib/testlib.js"></script>
</head>
<body>

        <div id="log">FAILED (This TC requires JavaScript enabled)</div>
        <script>log('inline script #1');
                var script = testlib.addScript('', { }, document.getElementsByTagName('body')[0], false);
                script.src='scripts/include-1.js';
                log('end script #1');
        </script>

        <script type="text/javascript">
        log( 'inline script #2' );
        var t = async_test()

        onload = function() {setTimeout(t.step_func(
          function() {
            assert_any(assert_array_equals, eventOrder, [['inline script #1', 'end script #1', 'external script #1', 'inline script #2'],
                                                         ['inline script #1', 'end script #1', 'inline script #2', 'external script #1']]);
            t.done()
          }),
        100)}
        </script>

</body></html>
